% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bgc_index.R
\docType{data}
\name{bgc_index}
\alias{bgc_index}
\title{Argo synthetic profile index}
\format{
## `bgc_index`
A data frame with 3,223 rows and 3 columns:
\describe{
  \item{parameters}{Parameters measured by the float}
  \item{parameter_data_mode}{Data processing mode (R: Raw, A: Adjusted, D: Delayed)}
  \item{wmo}{World Meteorological Object}
}
}
\source{
<ftp://ftp.ifremer.fr/ifremer/argo/argo_synthetic-profile_index.txt>
}
\usage{
bgc_index
}
\description{
Argo synthetic profile index
}
\keyword{datasets}
